refactor(conversation):WB-3957 create MessageItem component
